+++ config-templates/config-warning.php
@@ -1,0 +1,27 @@
+<?php
+
+/**
+ * This config template is able to show a warning to a user.
+ */
+$config = array(
+
+	/**
+	 * When true, the config file is switched on.
+	 */
+	'isOn' => true,
+
+	/**
+	 * When true, user is able to continue.
+	 */
+	'userCanContinue' => true,
+
+	/**
+	 * Title of the warning. It is possible to use HTML.
+	 */
+	'title' => '<b> Title of the warning </b>',
+
+	/**
+	 * Text of the warning. It is possible to use HTML.
+	 */
+	'text' => '<p> Due to maintenance it won\'t be possible to access our service from <b>1.1.2018 2:00</b> to <b>1.1.2018 4:00</b> </p>',
+);