+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 
 [Changed]
 - Updated README
+- Changed calls of RpcConnector methods (More information about changes in connectors and adapters in perun-simplesamlphp-module you can find in this [PR])
+
+[PR]:https://github.com/CESNET/perun-simplesamlphp-module/pull/34
 
 [Fixed]
 - Fixed duplicate property